adding routes and connecting with mysql table
