+++ server.js
@@ -1,27 +1,12 @@
-const mysql = require("mysql");
 const express = require("express");
 const bodyParser = require("body-parser");
+const mysqlConnection = require("./connection");
+const PeopleRoutes = require("./routes/people");
 
 var app = express();
 app.use(bodyParser.json());
 
-var mysqlConnection = mysql.createConnection({
-    host:"cakesdb1.cvnkqaqbljxc.us-east-1.rds.amazonaws.com",
-    user: "admin",
-    password : "Ireland1",
-    database : "cakesOclockSch", //schemaName
-    multipleStatements : true
-});
+app.use("/people", PeopleRoutes);
 
-mysqlConnection.connect((err)=>{
-    if(!err)
-    {
-        console.log("Connected");
-    }
-    else
-    {
-        console.log("Connection failed");
-    }
-})
 
 app.listen(3000);
